requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7
Pillow>=10.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
